backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
feedparser==6.0.10
redis==5.0.1
celery==5.3.4
openai==1.3.7
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
newspaper3k==0.2.8
nltk==3.8.1
textblob==0.17.1